project/requirements.txt
alembic==1.7.7
aniso8601==9.0.1
apt-clone==0.2.1
apturl==0.5.2
attrs==21.4.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.3
cloudinary==1.29.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
flasgger==0.9.5
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Autodoc==0.1.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.6.0
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
gnome-gmail==2.7
greenlet==1.1.2
grpcio==1.16.1
gunicorn==20.1.0
httplib2==0.14.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.0.3
jsonschema==4.6.1
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mistune==2.0.3
nemo-emblems==5.2.0
netaddr==0.7.19
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
onboard==1.4.1
openshot-qt==2.4.3
packaging==20.3
PAM==0.4.2
passlib==1.7.4
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.0.0
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.3
py==1.11.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.20.0
python-magic==0.4.16
python-xapp==2.2.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setproctitle==1.1.10
sib-api-v3-sdk==7.4.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
soupsieve==1.9.5
SQLAlchemy==1.4.36
systemd-python==234
tinycss2==1.0.2
tldextract==2.2.1
tomli==2.0.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.26.9
wadllib==1.3.3
webencodings==0.5.1
Werkzeug==2.1.2
WTForms==3.0.1
xkit==0.0.0
xlrd==1.1.0
youtube-dl==2021.4.26
zipp==3.8.0
